add notification

+++ app/models/concerns/mentionable.rb
@@ -8,35 +8,32 @@ module Mentionable
   private
 
   def parse_mention_from_body
-    receiver_ids = []
+    user_ids = []
 
     body.gsub!(/@(\w+)/) do |match|
-      if receiver = User.find_by_name($1)
-        unless receiver.id.eql?(user_id) || receiver_ids.include?(receiver.id)
-          receiver_ids << receiver.id
+      if user = User.find_by_name($1)
+        unless user_ids.include?(user.id)
+          user_ids << user.id
         end
       end
     end
 
-    create_new_mentions(receiver_ids)
+    create_new_mentions(user_ids)
 
-    delete_old_mentions(receiver_ids)
+    delete_old_mentions(user_ids)
   end
 
-  def create_new_mentions(receiver_ids)
-    receiver_ids.each do |receiver_id|
-      mentions.create(receiver_id: receiver_id, sender_id: user_id)
+  def create_new_mentions(user_ids)
+    user_ids.each do |user_id|
+      mentions.create(user_id: user_id)
     end
   end
 
-  def delete_old_mentions(receiver_ids)
-    if receiver_ids.present?
-      mentions.
-        where(sender_id: user_id).
-        where("receiver_id not in (#{receiver_ids.join(', ')})").
-        destroy_all
+  def delete_old_mentions(user_ids)
+    if user_ids.present?
+      mentions.where("user_id not in (#{user_ids.join(', ')})").destroy_all
     else
-      mentions.where(sender_id: user_id).destroy_all
+      mentions.destroy_all
     end
   end
 end
